fix: regression error from remove nginx dependency

+++ scripts/docker-init.sh
@@ -1,3 +1,15 @@
+if [ ! -f docker-dev ] && [ ! -f docker-release ]; then
+    echo "Internal Error: "
+    echo "This script should only be run in docker-dev or docker-release environment."
+    exit 1
+fi
+
+if [ ! -f scripts/.env ]; then
+    echo "Internal Error: "
+    echo "Missing scripts/.env file."
+    exit 1
+fi
+
 source scripts/.env set
 
 cp -r static-tmp/* static/
@@ -12,14 +24,18 @@ until pg_isready -h db -p 5432; do
 done
 
 if PGPASSWORD=${DB_PASSWORD} psql -U ntoj -d ntoj -h db -tAc "SELECT 1 FROM information_schema.tables WHERE table_name='challenge';" | grep -q 1; then
-    echo "exists"
+    echo "db exists"
 else
     sed -i "s/db_username/ntoj/g" ./scripts/oj.sql
     sed -i "s/db_name/ntoj/g" ./scripts/oj.sql
 
     ## Setup db
     PGPASSWORD=${DB_PASSWORD} psql -U ntoj -d ntoj -h db -f scripts/oj.sql
-    python3 scripts/add_admin.py ${ADMIN_NAME} ${ADMIN_PASSWORD} ${ADMIN_MAIL}
+    if [ -f docker-dev ]; then
+        poetry run python3 scripts/add_admin.py ${ADMIN_NAME} ${ADMIN_PASSWORD} ${ADMIN_MAIL}
+    else
+        python3 scripts/add_admin.py ${ADMIN_NAME} ${ADMIN_PASSWORD} ${ADMIN_MAIL}
+    fi
 fi
